+++ app/utils/ckb_utils.rb
@@ -1,7 +1,6 @@
 class CkbUtils
   def self.calculate_cell_min_capacity(output, data)
-    output.data = data if output.data.blank?
-    output.calculate_min_capacity
+    output.calculate_min_capacity(data)
   end
 
   def self.block_cell_consumed(transactions)
@@ -29,12 +28,13 @@ class CkbUtils
     return if cellbase.witnesses.blank?
 
     lock_script = generate_lock_script_from_cellbase(cellbase)
-    lock_script.to_hash(CkbSync::Api.instance)
+    lock_script.compute_hash
   end
 
   def self.generate_lock_script_from_cellbase(cellbase)
     witnesses_data = cellbase.witnesses.first.data
-    CKB::Types::Script.new(code_hash: witnesses_data.first[0..-3], args: [witnesses_data.last])
+    hash_type = witnesses_data.first[-2..-1] == "00" ? "data" : "type"
+    CKB::Types::Script.new(code_hash: witnesses_data.first[0..-3], args: [witnesses_data.last], hash_type: hash_type)
   end
 
   def self.generate_address(lock_script)
@@ -63,11 +63,12 @@ class CkbUtils
   end
 
   def self.block_reward(node_block_header)
-    primary_reward(node_block_header) + secondary_reward(node_block_header)
+    cellbase_output_capacity_details = CkbSync::Api.instance.get_cellbase_output_capacity_details(node_block_header.hash)
+    primary_reward(node_block_header, cellbase_output_capacity_details) + secondary_reward(node_block_header, cellbase_output_capacity_details)
   end
 
   def self.base_reward(block_number, epoch_number, cellbase = nil)
-    return cellbase.outputs.first.capacity.to_i if block_number.to_i == 0
+    return cellbase.outputs.first.capacity.to_i if block_number.to_i == 0 && cellbase.present?
 
     epoch_info = get_epoch_info(epoch_number)
     start_number = epoch_info.start_number.to_i
@@ -81,13 +82,11 @@ class CkbUtils
     end
   end
 
-  def self.primary_reward(node_block_header)
-    cellbase_output_capacity_details = CkbSync::Api.instance.get_cellbase_output_capacity_details(node_block_header.hash)
+  def self.primary_reward(node_block_header, cellbase_output_capacity_details)
     node_block_header.number.to_i != 0 ? cellbase_output_capacity_details.primary.to_i : 0
   end
 
-  def self.secondary_reward(node_block_header)
-    cellbase_output_capacity_details = CkbSync::Api.instance.get_cellbase_output_capacity_details(node_block_header.hash)
+  def self.secondary_reward(node_block_header, cellbase_output_capacity_details)
     node_block_header.number.to_i != 0 ? cellbase_output_capacity_details.secondary.to_i : 0
   end
 
@@ -127,12 +126,17 @@ class CkbUtils
     address_cell_consumed
   end
 
-  def self.update_block_reward_status!(current_block)
+  def self.update_block_reward!(current_block)
     target_block_number = current_block.target_block_number
     target_block = current_block.target_block
     return if target_block_number < 1 || target_block.blank?
 
-    target_block.update!(reward_status: "issued")
+    block_header = Struct.new(:hash, :number)
+    cellbase_output_capacity_details = CkbSync::Api.instance.get_cellbase_output_capacity_details(current_block.block_hash)
+    reward = CkbUtils.block_reward(block_header.new(current_block.block_hash, current_block.number))
+    primary_reward = CkbUtils.primary_reward(block_header.new(current_block.block_hash, current_block.number), cellbase_output_capacity_details)
+    secondary_reward = CkbUtils.secondary_reward(block_header.new(current_block.block_hash, current_block.number), cellbase_output_capacity_details)
+    target_block.update!(reward_status: "issued", reward: reward, primary_reward: primary_reward, secondary_reward: secondary_reward)
     current_block.update!(target_block_reward_status: "issued")
   end